+++ src/Network/constants.ts
@@ -2,8 +2,10 @@ import { NetworkConfig } from './types';
 // import { AVMConstants } from 'avalanche/dist/apis/avm';
 // import Avalanche, { AvalancheCore } from 'avalanche';
 import { Defaults } from 'avalanche/dist/utils';
+import { getRpcC, getRpcP, getRpcX } from './helpers/rpcFromConfig';
 
 export const MainnetConfig: NetworkConfig = {
+    rawUrl: 'https://api.avax.network',
     apiProtocol: 'https',
     apiIp: 'api.avax.network',
     apiPort: 443,
@@ -20,9 +22,17 @@ export const MainnetConfig: NetworkConfig = {
     evmChainID: Defaults.network[1]['C']['chainID'],
     // @ts-ignore
     avaxID: Defaults.network[1]['X']['avaxAssetID'],
+    get rpcUrl() {
+        return {
+            c: getRpcC(this),
+            p: getRpcP(this),
+            x: getRpcX(this),
+        };
+    },
 };
 
 export const TestnetConfig: NetworkConfig = {
+    rawUrl: 'https://api.avax-test.network',
     apiProtocol: 'https',
     apiIp: 'api.avax-test.network',
     apiPort: 443,
@@ -39,9 +49,17 @@ export const TestnetConfig: NetworkConfig = {
     evmChainID: Defaults.network[5]['C']['chainID'],
     // @ts-ignore
     avaxID: Defaults.network[5]['X']['avaxAssetID'],
+    get rpcUrl() {
+        return {
+            c: getRpcC(this),
+            p: getRpcP(this),
+            x: getRpcX(this),
+        };
+    },
 };
 
 export const LocalnetConfig: NetworkConfig = {
+    rawUrl: 'http://localhost:9650',
     apiProtocol: 'http',
     apiIp: 'localhost',
     apiPort: 9650,
@@ -56,6 +74,13 @@ export const LocalnetConfig: NetworkConfig = {
     evmChainID: Defaults.network[12345]['C']['chainID'],
     // @ts-ignore
     avaxID: Defaults.network[12345]['X']['avaxAssetID'],
+    get rpcUrl() {
+        return {
+            c: getRpcC(this),
+            p: getRpcP(this),
+            x: getRpcX(this),
+        };
+    },
 };
 
 // Default network connection