+++ src/Network/index.ts
@@ -1,12 +1,33 @@
 import { NetworkConfig } from './types';
-import { MainnetConfig } from '@/Network/constants';
-import { activeNetwork, setRpcNetwork, getEvmChainID } from '@/Network/network';
+import { MainnetConfig, TestnetConfig, LocalnetConfig } from '@/Network/constants';
+import { activeNetwork, setRpcNetwork, getEvmChainID, getConfigFromUrl, setRpcNetworkAsync } from '@/Network/network';
 import WebsocketProvider from '@/Network/providers/WebsocketProvider';
 import { bustErc20Cache } from '@/Asset/Erc20';
 
 export function setNetwork(conf: NetworkConfig) {
     setRpcNetwork(conf);
     bustErc20Cache();
+}
+
+/**
+ * Unlike `setNetwork` this function will fail if the network is not available.
+ * @param conf
+ */
+export async function setNetworkAsync(conf: NetworkConfig) {
+    await setRpcNetworkAsync(conf);
+    bustErc20Cache();
+}
+
+export function isFujiNetwork(activeNetwork: NetworkConfig) {
+    return activeNetwork.networkID === TestnetConfig.networkID;
+}
+
+export function isMainnetNetwork(activeNetwork: NetworkConfig) {
+    return activeNetwork.networkID === MainnetConfig.networkID;
+}
+
+export function isLocalNetwork(activeNetwork: NetworkConfig) {
+    return activeNetwork.networkID === LocalnetConfig.networkID;
 }
 
 // Default connection is Mainnet
@@ -16,4 +37,10 @@ export function getAvaxAssetID() {
     return activeNetwork.avaxID;
 }
 
-export { WebsocketProvider, getEvmChainID };
+export function getActiveNetworkConfig() {
+    return activeNetwork;
+}
+
+export { WebsocketProvider, getEvmChainID, getConfigFromUrl };
+
+export { NetworkConfig } from './types';