+++ src/helpers/utxo_helper.ts
@@ -2,8 +2,8 @@ import { UTXOSet as AVMUTXOSet } from 'avalanche/dist/apis/avm/utxos';
 import { UTXOSet as PlatformUTXOSet } from 'avalanche/dist/apis/platformvm/utxos';
 import { UTXOSet as EVMUTXOSet } from 'avalanche/dist/apis/evm/utxos';
 import { xChain, cChain, pChain } from '@/Network/network';
-import { BN } from 'avalanche';
 import { AvmImportChainType } from '@/Wallet/types';
+import { GetStakeResponse } from 'avalanche/dist/common';
 
 /**
  *
@@ -30,9 +30,6 @@ export async function avmGetAtomicUTXOs(addrs: string[], chainID: AvmImportChain
 
 // todo: Use end index to get ALL utxos
 export async function platformGetAtomicUTXOs(addrs: string[]): Promise<PlatformUTXOSet> {
-    // if (addrs.length > 1024) {
-    //     throw new Error('Number of addresses can not be greater than 1024.');
-    // }
     let selection = addrs.slice(0, 1024);
     let remaining = addrs.slice(1024);
 
@@ -56,10 +53,10 @@ async function evmGetAtomicUTXOs(addrs: string[]): Promise<EVMUTXOSet> {
     return result;
 }
 
-export async function getStakeForAddresses(addrs: string[]): Promise<BN> {
+export async function getStakeForAddresses(addrs: string[]): Promise<GetStakeResponse> {
     if (addrs.length <= 256) {
         let data = await pChain.getStake(addrs);
-        return data.staked;
+        return data;
     } else {
         //Break the list in to 1024 chunks
         let chunk = addrs.slice(0, 256);
@@ -67,8 +64,13 @@ export async function getStakeForAddresses(addrs: string[]): Promise<BN> {
 
         let chunkData = await pChain.getStake(chunk);
         let chunkStake = chunkData.staked;
+        let chunkUtxos = chunkData.stakedOutputs;
 
-        return chunkStake.add(await getStakeForAddresses(remainingChunk));
+        let next = await getStakeForAddresses(remainingChunk);
+        return {
+            staked: chunkStake.add(next.staked),
+            stakedOutputs: [...chunkUtxos, ...next.stakedOutputs],
+        };
     }
 }